requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.0
mysql-connector-python==8.1.0
sqlalchemy==2.0.23
pymysql==1.1.0
